refactor - Moved the Interfaces to their respective packages.
