+++ charts/src/main/java/dev/anirban/charts/linear/exceptions/DecorationMismatchStrategy.kt
@@ -1,7 +1,7 @@
 package dev.anirban.charts.linear.exceptions
 
 import dev.anirban.charts.linear.decoration.LinearDecoration
-import dev.anirban.charts.linear.interfaces.LinearDataStrategy
+import dev.anirban.charts.linear.data.LinearDataStrategy
 
 /**
  * This exception is thrown when the color array [LinearDecoration.plotPrimaryColor] and
