+++ charts/src/main/java/dev/anirban/charts/linear/plots/LinePlotStrategy.kt
@@ -4,8 +4,7 @@ import androidx.compose.ui.graphics.Path
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import androidx.compose.ui.graphics.drawscope.Stroke
 import dev.anirban.charts.linear.decoration.LinearDecoration
-import dev.anirban.charts.linear.interfaces.LinearDataStrategy
-import dev.anirban.charts.linear.interfaces.LinearPlotterStrategy
+import dev.anirban.charts.linear.data.LinearDataStrategy
 
 /**
  * This is the Line Plot class which implements the [LinearPlotterStrategy] Interface and makes a Line
